requirements.txt
fastapi
uvicorn
pymongo
elasticsearch
google-generativeai
